Changes after adding search engine

+++ src/graphComponents/Rating.java
@@ -5,7 +5,7 @@ public interface Rating {
 	public void setVote(String vote);
 	
 	public int getRating();
-	public void setRating(int rating);
+	public void setRating(double rating);
 	
 	public Movie getMovie();
 	public User getUser();
